Convert header config to TypeScript
There is no need for this to be a json, a TypeScript makes it easier to handle when the type is ESM.

Renaming the file to config, and moving the interface into file.
