--- src/components/header/header.json
@@ -1,6 +1,0 @@
-{
-	"moduleName": "ld-header",
-	"buttonId": "ld-menu-button",
-	"menuId": "ld-menu",
-	"expandedClassName": "expanded"
-}
